Move coral rot meta-data to "coral-size" hash. Alt col up/down.
In order to make sure the correct parts of the coral are available
as sites for new coral, alternate which spots are open based on
which col is offset up/down.

These calculations for coral rot were spread throughout the
code. So, moved them to a hash in the "state" -> "coral-size" hash.